requirements.txt
dj-database-url==0.5.0
Django==2.0.3
django-crispy-forms==1.7.2
gunicorn==19.7.1
pillow==4.0.0
psycopg2==2.7.1
whitenoise==3.3.1
scrapyd==1.2.0
scrapy==1.5.0
python-scrapyd-api==2.1.2